fixup! fixup! Merge branch 'master' into CHE-3621
