+++ plugins/plugin-ssh-machine/src/main/java/org/eclipse/che/plugin/machine/ssh/SshMachineFactory.java
@@ -10,19 +10,38 @@
  *******************************************************************************/
 package org.eclipse.che.plugin.machine.ssh;
 
+import com.jcraft.jsch.JSch;
+
 import org.eclipse.che.api.core.model.machine.Command;
 import org.eclipse.che.api.core.model.machine.Machine;
+import org.eclipse.che.api.core.model.machine.ServerConf;
 import org.eclipse.che.api.core.util.LineConsumer;
 import org.eclipse.che.api.machine.server.exception.MachineException;
+import org.eclipse.che.plugin.machine.ssh.jsch.JschSshClient;
 
+import javax.inject.Inject;
+import javax.inject.Named;
 import java.util.Map;
+import java.util.Set;
 
 /**
  * Provides ssh machine implementation instances.
  *
  * @author Alexander Garagatyi
+ * @author Max Shaposhnik
  */
-public interface SshMachineFactory {
+public class SshMachineFactory  {
+
+    private final int             connectionTimeoutMs;
+    private final Set<ServerConf> machinesServers;
+
+    @Inject
+    public SshMachineFactory(@Named("che.workspace.ssh_connection_timeout_ms") int connectionTimeoutMs,
+                             @Named("machine.ssh.machine_servers") Set<ServerConf> machinesServers) {
+        this.connectionTimeoutMs = connectionTimeoutMs;
+        this.machinesServers = machinesServers;
+    }
+
 
     /**
      * Creates {@link SshClient} to communicate with machine over SSH protocol.
@@ -32,8 +51,9 @@ public interface SshMachineFactory {
      * @param envVars
      *         environment variables that should be injected into machine
      */
-    SshClient createSshClient(SshMachineRecipe sshMachineRecipe,
-                              Map<String, String> envVars);
+    public SshClient createSshClient(SshMachineRecipe sshMachineRecipe, Map<String, String> envVars) {
+        return new JschSshClient(sshMachineRecipe, envVars, new JSch(), connectionTimeoutMs);
+    }
 
     /**
      * Creates ssh machine implementation instance.
@@ -43,9 +63,10 @@ public interface SshMachineFactory {
      * @param outputConsumer consumer of output from container main process
      * @throws MachineException if error occurs on creation of {@code Instance}
      */
-    SshMachineInstance createInstance(Machine machine,
-                                      SshClient sshClient,
-                                      LineConsumer outputConsumer) throws MachineException;
+    public SshMachineInstance createInstance(Machine machine, SshClient sshClient,
+                                             LineConsumer outputConsumer) throws MachineException {
+        return new SshMachineInstance(machine, sshClient, outputConsumer, this, machinesServers);
+    }
 
     /**
      * Creates ssh machine implementation of {@link SshMachineProcess}.
@@ -55,8 +76,8 @@ public interface SshMachineFactory {
      * @param pid virtual id of that process
      * @param sshClient client to communicate with machine
      */
-    SshMachineProcess createInstanceProcess(Command command,
-                                            String outputChannel,
-                                            int pid,
-                                            SshClient sshClient);
+    public SshMachineProcess createInstanceProcess(Command command, String outputChannel,
+                                                   int pid, SshClient sshClient) {
+        return new SshMachineProcess(command, outputChannel, pid, sshClient);
+    }
 }
