+++ plugins/plugin-ssh-machine/src/main/java/org/eclipse/che/plugin/machine/ssh/SshMachineModule.java
@@ -24,7 +24,7 @@ public class SshMachineModule extends AbstractModule {
     protected void configure() {
         bind(SshMachineInstanceProvider.class);
 
-        bind(SshMachineFactory.class).to(SshMachineFactoryImpl.class);
+        bind(SshMachineFactory.class).to(SshMachineFactory.class);
 
         bindConstant().annotatedWith(Names.named("machine.ssh.server.terminal.location")).to("~/che");
